unifiedrisk/common/data_fetcher.py
"""UnifiedRisk DataFetcher (网络版混合数据总线骨架).

说明：
- 优先使用 akshare 获取 A 股指数 / 成交额 / ETF / 两融等数据
- 预留东财（Eastmoney）接口位置，当前默认返回 0（避免报错）
- 所有外部依赖都放在函数内部 try/except ImportError 中，方便你按需安装

注意：
- 这里实现的是“可运行但偏保守”的版本：
    - 若 akshare 未安装 / 接口变更 / 网络异常 → 自动回退为 0，不影响主程序
    - 你可以在本文件基础上逐步替换为你自己的生产级抓取逻辑
"""

from __future__ import annotations

from typing import Dict, Any, Optional
from datetime import date, datetime


class DataFetcher:
    """数据抓取骨架（网络版）。

    当前版本仅实现 A 股日级数据：
    - fetch_ashare_daily_raw()

    返回的数据结构符合 UnifiedRisk AShareDailyEngine 所需 raw schema：
    {
        "ashare": {
            "index": {...},
            "fund_flow": {...},
            "liquidity": {...},
            "sentiment": {...},
            "style": {...},
            "valuation": {...},
        }
    }
    """

    # ========== 对外主入口 ==========

    def fetch_ashare_daily_raw(self, trade_date: Optional[date] = None) -> Dict[str, Any]:
        """抓取 A 股日级所需的全部数据，返回 UnifiedRisk raw schema。

        参数 trade_date:
            - None: 视为最新交易日（由各数据源自行判断）
            - date: 你可以在未来扩展为指定历史日期

        当前实现：
            - 以 "今日/最近一个交易日" 为目标
            - 若任一数据源失败 → 使用默认值 0
        """
        # 未来如需按日期获取，可在此构造 trade_date_str 传递给下游
        _ = trade_date or date.today()

        index_block = self._fetch_index_block()
        fund_flow_block = self._fetch_fund_flow_block()
        liquidity_block = self._fetch_liquidity_block(index_block, fund_flow_block)
        sentiment_block = self._fetch_sentiment_block()
        style_block = self._fetch_style_block()
        valuation_block = self._fetch_valuation_block()

        return {
            "ashare": {
                "index": index_block,
                "fund_flow": fund_flow_block,
                "liquidity": liquidity_block,
                "sentiment": sentiment_block,
                "style": style_block,
                "valuation": valuation_block,
            }
        }

    # ========== 各子模块抓取实现 ==========

    # ----- 1) 指数 + 成交额 -----

    def _fetch_index_block(self) -> Dict[str, Any]:
        """抓取上证 / 深成 / 创业板的日涨跌幅及成交额。

        优先使用 akshare 的现货接口，若失败则全部返回 0。
        """
        try:
            import akshare as ak
        except ImportError:
            return self._empty_index_block()

        try:
            # 使用 stock_zh_index_spot 获取当日实时行情（含涨跌幅、成交额等）
            df = ak.stock_zh_index_spot()
        except Exception:
            return self._empty_index_block()

        # 指数代码可能因 akshare 版本略有不同，这里使用常见代码：
        # 上证指数: 000001 / sh000001
        # 深证成指: 399001 / sz399001
        # 创业板指: 399006 / sz399006
        def _extract(code_candidates):
            for code in code_candidates:
                sub = df[df["代码"] == code]
                if not sub.empty:
                    row = sub.iloc[0]
                    # akshare 通常字段：涨跌幅、成交额（单位：亿元）
                    pct = float(row.get("涨跌幅", 0.0))
                    # 成交额字段名可能为 "成交额" 或 "成交额(亿)"
                    if "成交额(亿)" in row:
                        turnover = float(row.get("成交额(亿)", 0.0))
                    else:
                        turnover = float(row.get("成交额", 0.0))
                    return pct, turnover
            return 0.0, 0.0

        sh_pct, sh_turn = _extract(["000001", "sh000001"])
        sz_pct, sz_turn = _extract(["399001", "sz399001"])
        cyb_pct, cyb_turn = _extract(["399006", "sz399006"])

        return {
            "sh": {"pct": sh_pct, "turnover": sh_turn},
            "sz": {"pct": sz_pct, "turnover": sz_turn},
            "cyb": {"pct": cyb_pct, "turnover": cyb_turn},
        }

    def _empty_index_block(self) -> Dict[str, Any]:
        return {
            "sh": {"pct": 0.0, "turnover": 0.0},
            "sz": {"pct": 0.0, "turnover": 0.0},
            "cyb": {"pct": 0.0, "turnover": 0.0},
        }

    # ----- 2) 资金流向（北向替代 / 主力 / 两融） -----

    def _fetch_fund_flow_block(self) -> Dict[str, Any]:
        """抓取资金流向相关数据。

        - 北向替代：使用沪深 300 ETF（510300）与宽基 ETF（159919）的资金流入 proxy
        - 主力资金：预留东财接口，目前默认 0
        - 两融余额：若 akshare 可用则取近两日余额变化，否则 0
        """
        nb = self._fetch_northbound_proxy_via_etf()
        main_fund = self._fetch_main_fund_via_eastmoney()
        margin = self._fetch_margin_via_akshare()

        return {
            "northbound_proxy": nb,
            "main_fund": main_fund,
            "margin": margin,
        }

    def _fetch_northbound_proxy_via_etf(self) -> Dict[str, float]:
        """使用 ETF 成交额变化粗略构造北向替代因子。

        这里仅给出一个占位实现：
        - 若 akshare 可用，则尝试获取 510300 / 159919 的近期行情
        - 否则全部 0，后续你可以替换为更精细的逻辑
        """
        try:
            import akshare as ak
        except ImportError:
            return {
                "etf_510300_flow": 0.0,
                "etf_159919_flow": 0.0,
                "trend_3d": 0.0,
                "trend_5d": 0.0,
            }

        # 占位：仅返回 0，避免接口尚未调试完成时影响主流程
        # 你可以在此处接入：
        #   - ak.fund_etf_hist_em("510300", start=..., end=...)
        #   - 基于成交额和价格变化计算资金流入近 3/5 日趋势
        return {
            "etf_510300_flow": 0.0,
            "etf_159919_flow": 0.0,
            "trend_3d": 0.0,
            "trend_5d": 0.0,
        }

    def _fetch_main_fund_via_eastmoney(self) -> Dict[str, float]:
        """预留：使用东财接口抓取当日全市场主力资金净流入。

        当前版本为了稳健，默认返回 0，避免因接口变更导致报错。
        你可在此处自行补充 requests + 东财 API 的调用。
        """
        return {"inflow": 0.0}

    def _fetch_margin_via_akshare(self) -> Dict[str, float]:
        """使用 akshare 获取两融余额变化（占位实现）。"""
        try:
            import akshare as ak
        except ImportError:
            return {"change": 0.0}

        try:
            # akshare 提供的两融数据接口在不同版本下可能有差异，
            # 这里仅给出示例调用，你可根据实际返回结构做适配。
            df = ak.stock_margin_szse()  # 深市
            df2 = ak.stock_margin_sse()  # 沪市
        except Exception:
            return {"change": 0.0}

        # 占位：不做真实计算，仅返回 0，防止结构不兼容导致异常
        # 你可基于 df / df2 计算近两日余额变化百分比。
        return {"change": 0.0}

    # ----- 3) 流动性 -----

    def _fetch_liquidity_block(self,
                               index_block: Dict[str, Any],
                               fund_flow_block: Dict[str, Any]) -> Dict[str, Any]:
        """构造流动性因子所需的字段。

        - total_turnover: 直接使用指数部分三大指数成交额之和
        - turnover_ratio: 当前占位为 0（如有全市场换手率数据可填入）
        - northbound_etf_proxy: 使用 northbound_proxy 的某个字段（占位为 0）
        """
        sh_turn = float(index_block.get("sh", {}).get("turnover", 0.0))
        sz_turn = float(index_block.get("sz", {}).get("turnover", 0.0))
        cyb_turn = float(index_block.get("cyb", {}).get("turnover", 0.0))
        total_turnover = sh_turn + sz_turn + cyb_turn

        nb = fund_flow_block.get("northbound_proxy", {}) or {}
        nb_proxy = float(nb.get("etf_510300_flow", 0.0))

        return {
            "total_turnover": total_turnover,
            "turnover_ratio": 0.0,
            "northbound_etf_proxy": nb_proxy,
        }

    # ----- 4) 情绪 -----

    def _fetch_sentiment_block(self) -> Dict[str, Any]:
        """抓取情绪数据（涨停 / 跌停 / 炸板 / 高标）。

        当前版本仅返回占位 0，避免依赖复杂的网页抓取。
        你可以在此处接入：
            - 东方财富 涨停统计页面
            - 你的自建 scraper 结果
        """
        return {
            "limit_up": 0,
            "limit_down": 0,
            "open_limit_up": 0,
            "open_limit_up_success": 0,
            "high_leaders": {
                "max_continuous_limit": 0,
                "leading_sectors": [],
            },
        }

    # ----- 5) 风格轮动 -----

    def _fetch_style_block(self) -> Dict[str, Any]:
        """抓取风格与板块轮动数据。

        当前为占位实现：
        - 你可在此使用 akshare 获取：
            - 沪深 300 vs 中证 1000 / 创业板 指数，计算 large_small
            - 价值指数 vs 成长指数，计算 value_growth
            - 各行业指数涨跌幅，填入 sector_strength
        """
        return {
            "large_small": 0.0,
            "value_growth": 0.0,
            "sector_strength": {},
        }

    # ----- 6) 估值 -----

    def _fetch_valuation_block(self) -> Dict[str, Any]:
        """抓取估值与筹码相关数据。

        当前为占位实现：
        - 你可选用：
            - 自己维护的 A 股全市场估值 CSV
            - akshare 的指数估值接口
            - 东方财富的估值统计
        """
        return {
            "index_valuation": {
                "sh_pe_percentile": 0.0,
                "sz_pe_percentile": 0.0,
                "cyb_pe_percentile": 0.0,
            },
            "market_pb": 0.0,
            "concentration": 0.0,
        }
